Add ENOBUFS handling for unsolicited messages
This can happen when large burst of messages come all of a sudden, which
happen very easily when routing protocols are involved (e.g. BGP). The
current implementation incorrectly assumes that any failure to read from
the socket is akin to the socket closed. This is not the case.

This commit adds handling for this specific error by generating a
`NetlinkPayload::Overrun(_)` message that users receive on their
unsolicited message channel. Since this is just an additional message,
there is no breaking change for existing users and they are free to
ignore it if they do not want to handle it, or handle it by e.g. resyncing.