Check for closed unsolicited messages channel
The unsolicited messages channel is used to forward
unsolicited messages like multicast messages to the user.

However, if none are sent by the kernel, the connection
stays active even if the rx side of the channel was already
dropped. This finally leads to resource exhaustion (e.g. too
many open file descriptors) if new connections are repeatedly
opened.

Therefore, explicitly check if the channel is already closed
and if so, take the tx side to allow for shutdown of the connection.

Signed-off-by: Florian Kauer <florian.kauer@linutronix.de>